ML Services/luggage-tracking-ml/requirements.txt
Flask==2.3.0
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.0.0
gunicorn==21.2.0
joblib==1.3.0
requests==2.31.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
